+++ templates/dummy_impl_prologue.cc
@@ -37,12 +37,12 @@ void dummyClassFinalizer(KNativePointer* ptr) {
 namespace OHOS::Ace::NG {
 namespace Bridge {
 Ark_NodeHandle CreateNode(GENERATED_Ark_NodeType type, Ark_Int32 id, Ark_Int32 flags) {
-    static long peer_num = 1;
+    static uintptr_t peer_num = 1;
     Ark_NodeHandle result = (Ark_NodeHandle) peer_num++;
 
     if (needGroupedLog(2)) {
         std::string _logData;
-        _logData.append("  Ark_NodeHandle peer" + std::to_string((long)result) + " = GetBasicNodeApi()->createNode(GENERATED_Ark_NodeType("
+        _logData.append("  Ark_NodeHandle peer" + std::to_string((uintptr_t)result) + " = GetBasicNodeApi()->createNode(GENERATED_Ark_NodeType("
             + std::to_string(type) + "), " + std::to_string(id) + ", " + std::to_string(flags) + ");\n");
         appendGroupedLog(2, _logData);
     }
